Workaround broken extern weak hidden handling
The version of GCC that gets built as part of Yocto 1.6 has a bug in
handling symbols that use __attribute__((weak, visibility("hidden")))
and are at the same time declared external.

The .hidden assembler directive gets accidentally omitted (on ARM).

Unfortunately lttng-ust makes heavy use of this combination. The source
file that emits this symbols is modified to enforce emitting .hidden
directives for the symbols.